[Refactor] Extract fuse search class as a separate file
